BRFSS Caregiver (CG) Module Data

BRFSS Caregiving Module Questions asked from 2015-2019 https://www.cdc.gov/aging/healthybrain/brfss-faq-caregiver.htm.

brfss_cg

Format

A "long" data frame with XX rows of individual BRFSS respondents in a given state and year and YY column variables:

cg_d_num

indicator, numeric variable of whether respondent was Caregiver (cg_d_num=1) or not (cg_d_num=0)

cg_d_fct

indicator, factor variable of whether respondent was Caregiver (cg_d_fct="CG") or not (cg_d_fct="Non-CG")

state

state FIPS code, labeled with state alphabetic abbreviation

seqno

identificaiton variable

year

Numeric year, 2014-2019

sgm_wt_raw

Original, raw sampling weight: 2014-2019

Source

BRFSS Annual Survey Data https://www.cdc.gov/brfss/annual_data/annual_data.htm.

Examples

# To adjust the sampling weight (cg_wt_adj) by dividing the
# sampling weight (cg_wt_raw) by the number of instances
# a state is in the data, run:
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── tidyverse 1.3.0 ──

#> ✓ ggplot2 3.3.2     ✓ purrr   0.3.4
#> ✓ tibble  3.0.3     ✓ dplyr   1.0.2
#> ✓ tidyr   1.1.2     ✓ stringr 1.4.0
#> ✓ readr   1.3.1     ✓ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── tidyverse_conflicts() ──
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

data(brfss_cg)
waves<-brfss_cg %>%
  filter(year %in% 2016:2018) %>% #keeping only 2016-2018 for illustration
  group_by(year,state) %>%
  slice(1) %>% #keeping the first observation of each state + year
  ungroup() %>%
  group_by(state) %>% #grouping by state
  count() %>% #counting how many years the state was included
  rename(wave=n) #renaming as wave
brfss_cg<-full_join(brfss_cg,waves,by="state") %>%
  mutate(cg_wt_adj = cg_wt_raw/wave) #adjusting the weight by number of waves