BRFSS Demographic, Socioeconomic, Health Care, Status and Behaviors Data

BRFSS Demographic, Socioeconomic, Health Care, Status and Behaviors Data

brfss_core

Format

A long data frame with rows of individual BRFSS respondents in a given state and year and the following column variables:

age_num

numeric, quantiative 18-99

age_cat

factor, categorical 5-year age categories (18-24, 25-29,…95-99)

age_10ycat_fct

factor, categorical 10-year age categories (18-24, 25-34,35-44,45-54,55-64,65+)

millennial_d_num

numeric, dichotomous 1/0

millennial_d_fct

factor, dichotomous Yes/No

agege45_d_num

numeric, dichotomous 1/0

agege45_d_fct

factor, dichotomous Yes/No

agege65_d_num

numeric, dichotomous 1/0

agege65_d_fct

factor, dichotomous Yes/No

mstat_cat_fct

factor, categorical Married, Divorced, Widowed, Separated, Never Married, Coupled

mstat_cat_fct

factor, categorical Married, Divorced, Widowed, Separated, Never Married, Coupled

married_d_fct

factor, dichotomous: Married or Coupled Yes/No (No: divorced, widowed, separated, never married)

chld_num

numeric, quantiative 0-87

chld_cat_fct

factor, categorical None, One, Two or More

kids_d_fct

factor, dichotomous Any Kids Yes/No

race_cat_fct

factor, categorical White only, Black or African American only, American Indian or Alaskan Native only,Asian only, Native Hawaiian or other Pacific Islander only,Other race only,Multiracial

]
ethn_cat_fct

factor, categorical Hispanic, Non-Hispanic

raceth_cat_fct

factor, categorical White Non-Hispanic; Black, Non-Hispanic; Other, Non-Hispanic; Multiple, Non-Hispanic; Hispanic

reth_imp_cat_fct

factor, categorical from imputed race (no missing): White Non-Hispanic; Black, Non-Hispanic; Other, Non-Hispanic; Asian, Non-Hispanic; Native, Non-Hispanic, Hispanic

sex_d_fct

factor, dichotomous female/male

fem_d_num

numeric, dichotomous 0/1 (male/female)

vtrn_d_num

numeric, dichotomous 0/1 (Non-Vet/Veteran)

vtrn_d_fct

factor, dichotomous Non-Veteran, Veteran

msa_cat_fct

factor, categorical City_Center, City_County, Suburb, Outside MSA

msa_d_fct

factor, dichotomous Yes/No

metro_d_fct

factor, dichotomous Yes/No: 2018-2019 only (raw BRFSS variable: _METSTAT)

urban_d_fct

factor, dichotomous Yes/No: 2018-2019 only (raw BRFSS variable: _URBSTAT)

drnkbng_d_fct

factor, dichotomous Yes/No

drnkhvy_d_fct

factor, dichotomous Yes/No

bmi_cat_fct

factor, categorical Underweight, Normal Weight, Overweight, Obese, Unkown

bmi_num

numeric, quantitative value of BMI

bmi_40d_fct

BMI>=40: factor, dichotomous (No/Yes)

obese_d_fct

BMI>=30: factor, dichotomous (No/Yes)

hiv_d_num

numeric, dichotomous 0/1 (No/Yes)

hiv_d_fct

factor, dichotomous Yes/No

fluvac_d_fct

factor, dichotomous Yes/No

fluvac_d_num

numeric, dichotomous 0/1 (No/Yes)

ltpa_d_fct

factor, dichotomous Yes/No

smk_cat_fct

factor, categorical Current smoker, Former smoker, Never smoker

smoker_d_fct

factor, categorical: Current smoker vs. Former/Never Yes/No

drcost_d_fct

factor, dichotomous Yes/No

hcplan_d_fct

factor, dichotomous Yes/No

chckup_cat_fct

factor, categorical <1 year, 1-2 years, 2-5 years, 5+ years, Never

dntst_cat_fct

factor, categorical <1 year, 1-2 years, 2-5 years, 5+ years, Never

chron_num

numeric, quantiative 0-7

chron_d_fct

factor, dichotomous Yes/No

cvd_d_num

Cardiovascular Disease: numeric, dichotomous 0/1 (No/Yes)

strk_d_num

Stroke: numeric, dichotomous 0/1 (No/Yes)

diab_d_num

Diabetes: numeric, dichotomous 0/1 (No/Yes)

asth_d_num

Asthma: numeric, dichotomous 0/1 (No/Yes)

arth_d_num

Arthritis: numeric, dichotomous 0/1 (No/Yes)

copd_d_num

COPD: numeric, dichotomous 0/1 (No/Yes)

cncr_d_num

Cancer: numeric, dichotomous 0/1 (No/Yes)

kidn_d_num

Kidney Disease: numeric, dichotomous 0/1 (No/Yes)

dep_d_fct

factor, dichotomous No/Yes

mentqol_num

numeric, quantitative, 0-30

mentqol14_d_num

numeric, dichotomous 0/1 (No/Yes)

mentqol14_d_fct

factor, dichotomous No/Yes

physqol_num

numeric, quantitative, 0-30

physqol14_d_num

numeric, dichotomous 0/1 (No/Yes)

physqol14_d_fct

factor, dichotomous No/Yes

srh_d_fct

factor, dichotomous: Good+ vs. Fair/Poor

srh_cat_fct

factor, categorical: Excellent, Very Good, Good, Fair, Poor

empl_cat_fct

factor, categorical: Out of work, Employed for wages, Self-employed,A homemaker, A student, Retired, Unable to work

employed_d_fct

factor, dichotomous: Employed for Wages or Self-Employed? Yes / No

inc_cat_fct

factor, categorical: <$10,000, $10,000-$14,999, $15,000-19,999, $20,000-$24,999,$25,000-34,999, $35,000-$49,999,$50,000-74,999, $75,000+,Don't know or refused

income_4cats_fct

factor, categorical: <$20,000; 20,000-34,999; 35,000-74,999; $75,000+

educ_cat_fct

factor, categorical:<High School,High School,Some College,College or More

college_d_fct

factor, categorical: College or More Yes/No

x_state

numeric variable of state

state

state FIPS code, labeled with state alphabetic abbreviation

fips

state FIPS code, numeric

seqno

identificaiton variable

year

Numeric year: 2014-2019

x_psu

Primary Sampling Unit Variable

x_ststr

Sampling Strata Variable

var_wt_raw

Original,raw sampling weight: 2014-2020

version_var

Character, Version of data: Core (X_LLCPWT), V1 (X_LCPWTV1), V2(X_LCPWTV2), V3 (X_LCPWTV3)

Source

BRFSS Annual Survey Data https://www.cdc.gov/brfss/annual_data/annual_data.htm.

Examples

# To adjust the sampling weight (var_wt_raw) by dividing
# the sampling weight by the number of instances a state
# is in the data, run:
library(tidyverse)
data(brfss_core)
waves<-brfss_core %>%
  filter(year %in% 2016:2018) %>% #keeping only 2016-2018 for illustration
  group_by(year,state) %>%
  slice(1) %>% #keeping the first observation of each state + year
  ungroup() %>%
  group_by(state) %>% #grouping by state
  count() %>% #counting how many years the state was included
  rename(wave=n) #renaming as wave
brfss_core<-full_join(brfss_core,waves,by="state") %>%
  mutate(var_wt_adj = var_wt_raw/wave) #adjusting the weight by number of waves